Remove line items and total from CheckoutProvider
This removes the useTotal function exported by the
`@automattic/composite-checkout` package as well as the whole concept of
line items within the package. Since the demo and the tests depend on
that concept (having no line item rendering of their own), this change
moves the necessary components to the tests and demo.